add MapViewModel to handle location updates
